lando pass vm_flavor based on job and combine redundant test utils
